Decouple DrQ agent.
* Add a separate learner for DrQ agent for easy reuse.
* Fix an issue in types for the Gym subproc wrapper.
